refactor(raw): split timer store from actor context
